:art: NTG-45 enhance: code solid design
