+++ src/main/java/com/alessandragodoy/customerms/repository/CustomerRepository.java
@@ -1,6 +1,6 @@
 package com.alessandragodoy.customerms.repository;
 
-import com.alessandragodoy.customerms.model.entity.Customer;
+import com.alessandragodoy.customerms.model.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
